backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nodeenv==1.8.0
numpy==1.26.0
pandas==2.1.1
prisma==0.10.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
